backend/requirements.txt
Flask==2.1.4
Flask-CORS==3.1.1
PyJWT==2.1.0
